Changed MOD from long to int (noticed drastic performance difference when taking modulos of ints with a long, versus taking modulos of ints with an int)

+++ src/baobab/B.java
@@ -37,7 +37,7 @@ public class B {
 
         /************************** UTILITY CODE BELOW THIS LINE **************************/
 
-        long MOD = (long)1e9 + 7;
+        int MOD = (int)1e9 + 7;
 
         boolean closeToZero(double v) {
             // Check if double is close to zero, considering precision issues.
